Be a bit more restrictive in the negative lookahead for properties
Fixes #75 and #115

There is a bit of negative lookahead for properties to avoid matching nested types
as property declarations. Consider the following code:

```C#
class Outer
{
    class Inner
    {
    }
}
```

When the line "class Inner" is matched it really looks like a property declaration
where the type is "class" and the name is "Inner". So, there's negative lookahead to
ensure that properties can't start with "class" or other keywords.

Before this change, the lookahead was matching `.*(class|event|etc...)`, which was
too permissive and allowed it to eat up comments. This change adjusts the lookahead to
`[[:word:][:space:]]*(class|event|etc...)` to make it a little less restrictive.

In addition, there was a spurious `\s*` after the negative lookahead that seems like it
was only there to help distinguish field and property declarations. This is more properly
handled by just ordering the rules differently (putting properties first) and removing
the `\s*`.